Fix ect to etc...
Is viewMenue a redundancy? Also, hospitalOptionMenue? Some of the options are not implemented, tried adding a patient and said query failed before asking me to input information. There should probably be format changes as well in code, for example I personally think it is ugly to have db.add("orders insert",br) in a large if else. I think it can be reduced to (if each option is unique) a container like it was before and instead of add being add(string, string) it could be add(int, string) or the keys could be changed. For instance, I think it would look good if it were the ____whatever____ in the designated class. Then we can do something like add(4,br) and it would spit out the statement, separating logic code from format code. I am not sure in java if it matters too much but I think it would be better to get rid of boolean t = true and set to false later when it can be handled with a break, I know it is better in c languages(not instantiating a value that is not needed). In fact what might be best is making an abstract menu class with the other menus being subsets of menu, and if needed can be composites. Overall I do like this menu layout more than my janky layout lol. What does everyone else think?